Create evict.h and move structs/macros in it
