+++ src/include/cache.h
@@ -8,46 +8,6 @@
 
 #pragma once
 
-/*
- * Helper: in order to read without any calls to eviction, we have to ignore the cache size and
- * disable splits.
- */
-#define WT_READ_NO_EVICT (WT_READ_IGNORE_CACHE_SIZE | WT_READ_NO_SPLIT)
-
-/*
- * Tuning constants: I hesitate to call this tuning, but we want to review some number of pages from
- * each file's in-memory tree for each page we evict.
- */
-#define WT_EVICT_MAX_TREES WT_THOUSAND /* Maximum walk points */
-#define WT_EVICT_WALK_BASE 300         /* Pages tracked across file visits */
-#define WT_EVICT_WALK_INCR 100         /* Pages added each walk */
-
-/*
- * WT_EVICT_ENTRY --
- *	Encapsulation of an eviction candidate.
- */
-struct __wt_evict_entry {
-    WT_BTREE *btree; /* Enclosing btree object */
-    WT_REF *ref;     /* Page to flush/evict */
-    uint64_t score;  /* Relative eviction priority */
-};
-
-#define WT_EVICT_QUEUE_MAX 3    /* Two ordinary queues plus urgent */
-#define WT_EVICT_URGENT_QUEUE 2 /* Urgent queue index */
-
-/*
- * WT_EVICT_QUEUE --
- *	Encapsulation of an eviction candidate queue.
- */
-struct __wt_evict_queue {
-    WT_SPINLOCK evict_lock;                /* Eviction LRU queue */
-    WT_EVICT_ENTRY *evict_queue;           /* LRU pages being tracked */
-    WT_EVICT_ENTRY *evict_current;         /* LRU current page to be evicted */
-    uint32_t evict_candidates;             /* LRU list pages to evict */
-    uint32_t evict_entries;                /* LRU entries in the queue */
-    wt_shared volatile uint32_t evict_max; /* LRU maximum eviction slot used */
-};
-
 /* Cache operations. */
 typedef enum __wt_cache_op {
     WT_SYNC_CHECKPOINT,
@@ -241,11 +201,6 @@ struct __wt_cache {
     uint32_t flags;
 };
 
-#define WT_WITH_PASS_LOCK(session, op)                                                   \
-    do {                                                                                 \
-        WT_WITH_LOCK_WAIT(session, &cache->evict_pass_lock, WT_SESSION_LOCKED_PASS, op); \
-    } while (0)
-
 /*
  * WT_CACHE_POOL --
  *	A structure that represents a shared cache.
@@ -275,10 +230,3 @@ struct __wt_cache_pool {
  * file.
  */
 #define WT_IS_HS(dh) F_ISSET(dh, WT_DHANDLE_HS)
-
-/* Flags used with __wt_evict */
-/* AUTOMATIC FLAG VALUE GENERATION START 0 */
-#define WT_EVICT_CALL_CLOSING 0x1u  /* Closing connection or tree */
-#define WT_EVICT_CALL_NO_SPLIT 0x2u /* Splits not allowed */
-#define WT_EVICT_CALL_URGENT 0x4u   /* Urgent eviction */
-/* AUTOMATIC FLAG VALUE GENERATION STOP 32 */
