+++ src/include/wt_internal.h
@@ -516,6 +516,7 @@ typedef uint64_t wt_timestamp_t;
 #include "btmem.h"
 #include "btree.h"
 #include "cache.h"
+#include "evict.h"
 #include "capacity.h"
 #include "cell.h"
 #include "checkpoint.h"
